Moved vehicle.py to control and changed imports
