+++ anki/__init__.py
@@ -1,4 +1,4 @@
-from .vehicle import Vehicle, Lights
+from .control.vehicle import Vehicle, Lights
 from .utility.track_pieces import TrackPiece, TrackPieceTypes
 from .control.controller import Controller
 from .utility.lanes import Lane3, Lane4
